Jittery Sequencer Animations with RateScale > 0.0
Two changes fixed this. First we invert out the RateScale when calculating the PlayRate inside  FAnimNode_SequenceEvaluator::UpdateAssetPlayer, since it get's multiplied back in the CreateTickRecordForNode function.   This stops the jittering but the animation playback isn't scaled. To fix that we modify the Sequencer sections PlayRate by the asset'ss RateScale.

#jira UE-63636

#rb max.chen, martin.wilson

[CL 4867670 by Mike Zyracki in Dev-Sequencer branch]