Fix serialization issues with MeshDescription.
FMeshDescriptionBulkData is serialized directly to a package, but its bulk data contains a compressed FMeshDescription which is serialized at a later stage on demand. It is mandatory that the versioning for each of these stay in sync, otherwise MeshDescription will fail to load following a version bump.

This fix ensures that, when saving FMeshDescriptionBulkData, its bulk data is reserialized with the latest versioning. It is not always the case that the mesh description would be otherwise updated prior to saving the FMeshDescriptionBulkData.

This fix also ensures that the internal state of FMeshDescriptionBulkData is saved and restored around an undo/redo. Previously this could get lost when performing an undo in the static mesh editor.

#rb Ryan.Schmidt
#jira none

#ROBOMERGE-SOURCE: CL 4929630 in //UE4/Release-4.22/...
#ROBOMERGE-BOT: RELEASE (Release-4.22 -> Main)

[CL 4930026 by richard talbotwatkin in Main branch]