Jittery Sequencer Animations with RateScale > 0.0
Two changes fixed this. First we invert out the RateScale when calculating the PlayRate inside  FAnimNode_SequenceEvaluator::UpdateAssetPlayer, since it get's multiplied back in the CreateTickRecordForNode function.   This stops the jittering but the animation playback isn't scaled. To fix that we modify the Sequencer sections PlayRate by the asset'ss RateScale.

#jira UE-63636
#rb max.chen, martin.wilson

#ROBOMERGE-OWNER: robert.manuszewski
#ROBOMERGE-AUTHOR: max.chen
#ROBOMERGE-SOURCE: CL 4896912 in //UE4/Release-4.22/... via CL 4896913
#ROBOMERGE-BOT: CORE (Main -> Dev-Core)

[CL 4947075 by max chen in Dev-Core branch]