Layering Support for Channel Key Editing via IKeyArea.
With the previous work here we supporting Layering with external values correctly working with the TrackEditors which work when setting keys via the details panel, auto key, or the hot keys. However if you pressed the key button icon next to the channel in the sequencer UI it would only now always just set the current value which could be different than what it should be based upon the external value (like if you moved the sphere in the jira ticket).

So we added support for an extra OnGetCurrentValueAndWeight to the TMovieSceneExternalValue in order to modify the value based upone the global diff/weight as we did before. This specialization only happens for <float> and currently only for the Transform/Vector sections (other like ComposurePostMovieSettings seemed margiinally useful and we don't support ExternalValues for generic float chanels for some reason).

#jira UE-69648
#rb max.chen

#ROBOMERGE-OWNER: robert.manuszewski
#ROBOMERGE-AUTHOR: mike.zyracki
#ROBOMERGE-SOURCE: CL 4997298 in //UE4/Release-4.22/... via CL 4997300
#ROBOMERGE-BOT: CORE (Main -> Dev-Core)

[CL 5060083 by mike zyracki in Dev-Core branch]