Fixed issue where a "Draw Material to RenderTarget" node invoked in the first frame of an application would result in the render target being immediately cleared afterwards.
In the case where a blueprint is invoked on the first frame to render a material into a render target, the render target can be simultaneously queued for a deferred update. After rendering the material to the render target, the actual scene is rendered, which flushes this deferred update queue. The render target is then immediately cleared. This change flushed the queue prior to rendering a canvas to a render target, because we need all deferred clears to occur beforehand.

#jira UE-36558
#rb Rolando.Caloca

[CL 5064828 by zach bethel in Dev-Rendering branch]