Allow game localization targets to gather engine sources from an out-of-source project
The localization pipeline locks the relative paths used by a target to either the engine or project root (depending on the target type), however the UE4 engine isn't localized into many languages, so project targets often have to override some of the engine localization to support their additional languages. For in-source builds, this was possible by using a project relative path to the engine directory (which was in a fixed location), however for any out-of-source projects (such as those in the "Unreal Projects" directory) there was no way to do this in a portable way.

This change adds a "path root" to every include and exclude path, which can be set to "Auto" (the default, which detects the root based on the target type), "Engine" (to force the path to be resolved relative to the engine root), or "Project" (to force the path to be resolved relative to the project root). This allows out-of-source projects to specify paths relative to the engine in a portable way.

#jira UE-70296
#rb Lauren.Ridge

#ROBOMERGE-OWNER: lina.halper
#ROBOMERGE-AUTHOR: jamie.dale
#ROBOMERGE-SOURCE: CL 5073204 in //UE4/Release-4.22/... via CL 5073205
#ROBOMERGE-BOT: ANIM (Main -> Dev-Anim)

[CL 5137242 by jamie dale in Dev-Anim branch]