Fixed issue where MaterialParameterCollections would sometimes revert to default values in PIE.
When invoking PIE, the current UWorld is duplicated, along with all actor instances. Material Parameter Collections have a unique instance for each world. The current material parameter data associated with the source world was not properly serialized during duplication, and the construction script for the actor was not run, therefore the parameter values would revert to their defaults.

This change makes the instance data serialize out during a duplication, but only for PIE mode.

#jira UE-69616
#rb Daniel.Wright

#ROBOMERGE-OWNER: ryan.vance
#ROBOMERGE-AUTHOR: zach.bethel
#ROBOMERGE-SOURCE: CL 5070596 in //UE4/Release-4.22/... via CL 5070602
#ROBOMERGE-BOT: DEVVR (Main -> Dev-VR)

[CL 5140965 by zach bethel in Dev-VR branch]