Added support for connecting to an existing editor instance to run Python commands on it remotely
This is off by default, and discovery is limited to the local host by default for safety. To enable it you need to turn on remote discovery in the Python plug-in project settings, which will make your editor discoverable for remote execution (via external scripts using the remote_execution.py module).

Remote execution uses a protocol of UTF-8 encoded JSON, and a combination of UDP-based multicast for discovery and remote connection requests, and TCP-based peer-to-peer communication for the actual command channel. See PythonScriptRemoteExecution.cpp for the full protocol definition.

remote_execution.py serves as both a reference implementation for the protocol implementation needed to connect to an editor instance, and as an out-of-the-box solution for applications that are written in (or can execute) Python (by importing it as a module and creating a RemoteExecution instance). remote_execution.py also provides a minimal example at the bottom of the file that demonstrates how to use the API, and this demo application can be seen by executing the file directly.

[FYI] Aaron.Carlisle
#rb Chris.Gagnon

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: jamie.dale
#ROBOMERGE-SOURCE: CL 5229289 via CL 5236106 via CL 5236230
#ROBOMERGE-BOT: ENGINE (Main -> Dev-Networking)

[CL 5254936 by jamie dale in Dev-Networking branch]