#jira UE-70329 - Name given to a component on creation was not preserved in multi-user context.
The editor assumed that renaming a component on creation was part of the creation process (from the undo perspective) and was not made part of any transaction so that undoing after giving the initial name would actually undo the creation. For multi-user, giving a new name, on creation, must be a transaction so that it get recorded and replicated. Fixed the issue by keeping the create transaction on-going until the initial name is given. The rename transaction gets folded under any on-going create. From the undo/redo buffer, this is still a single transaction, but now, the rename gets replicated in the multi-user context.

#rb Jamie.Dale

#ROBOMERGE-OWNER: robert.manuszewski
#ROBOMERGE-AUTHOR: patrick.laflamme
#ROBOMERGE-SOURCE: CL 5339730 in //UE4/Release-4.22/... via CL 5339744
#ROBOMERGE-BOT: CORE (Main -> Dev-Core)

[CL 5407738 by patrick laflamme in Dev-Core branch]