#jira UE-69820 Motion Controller meshes do not track with Motion Controller position within VRMode
This issue was introduced by the fix that allowed MotionControllerComponents to correctly replicate in networked games.  Previously the local motion controllers overrode the remote controller's positions.  This was fixed by only updating if HasLocalNetOwner was true.  But that was returning false for VREditorAvatarActor, as it's Owner was null.
So for now, VREditorAvatarActor's HasLocalNetOwner now always returns true.  When using Multi-User with two VRMode editors, we might have to do something else here for when this is called by MotionControllerComponent::PollControllerState()

#rb Lauren.Ridge

#ROBOMERGE-OWNER: robert.manuszewski
#ROBOMERGE-AUTHOR: joe.conley
#ROBOMERGE-SOURCE: CL 5391928 in //UE4/Release-4.22/... via CL 5391930
#ROBOMERGE-BOT: CORE (Main -> Dev-Core)

[CL 5444569 by joe conley in Dev-Core branch]