#jira UE-71665 [CrashReport] UE4Editor-Kismet!SSCSEditor::AddNewComponent(UClass *,UObject *,bool,bool) [SSCSEditor.cpp:5121]
Ensured that the 'create + rename' transaction was ended if the component name editable text failed to scroll into view during the frame the component was created. The previous code assumed the name would always be scrolled into view with the focus to let the user rename the newly created component, then close the 'create + rename' transaction as a single 'undo' step, but if the widget did not have any space to display the item, the operation would be postponed until the UI gets resized to display it, leaving the transaction unexpectedly open for an undefined amount of frames (messing the undo/redo buffer). The new code detects if the component was scrolled into view during the frame and if not, the transaction is closed. In such case, the user has to rezise the UI to bring the component into view and rename it.

#rb Jamie.Dale

#ROBOMERGE-SOURCE: CL 5456550 in //UE4/Release-4.22/...
#ROBOMERGE-BOT: RELEASE (Release-4.22 -> Main)

[CL 5456593 by patrick laflamme in Main branch]