Fixed sky diffuse being invisible in project with movable skylight and r.SupportStationarySkylight is false. Behavior was changed in 4691984.
SupportStationarySkylight should not be used to conditionally create a sky or not. Sky is always used for default environment diffuse (just can be rendered in non base pass but environment pass depending on whether lighting/sky is static/stationary of movable).
A SphereReflectionCapture is not enough to get diffuse light (it can only be used for specular). Which is what we see in this JIRA bug.

Fix is to always create the sky and just make it visible or not (the new behavior added in 4691984 and that is correct)

#rb Jurre.deBaare


#ROBOMERGE-SOURCE: CL 5683666 via CL 5683884

[CL 5683886 by sebastien hillaire in Main branch]