#jira UE-72316 - Crash when duplicating multiple components at once
An assertion fired up when SSCSEditor::AddNewComponent() was called in a loop (or multiple times in the same frame) with every component scheduled to get the focus. This unexpectedly activated the 'create + rename' transaction mechansim multiple times but only one item (the last in the frame) would end up being focused, scrolled into view and renamable on creation. Instead of asserting, if an ongoing transaction is ongoing, assumes it is because AddNewComponet() is being called in a loop, close the previous transaction and open a new one, expecting the last component added in the frame (with focus request) to close its own transaction once scrolled into view and edited by user.

#rb Marc.Audy
#lockdown Nick.Penwarden

#ROBOMERGE-OWNER: lina.halper
#ROBOMERGE-AUTHOR: patrick.laflamme
#ROBOMERGE-SOURCE: CL 5790810 in //UE4/Release-4.22/... via CL 5790816
#ROBOMERGE-BOT: ANIM (Main -> Dev-Anim)

[CL 5796408 by patrick laflamme in Dev-Anim branch]