Fixed issue where the provisioning profile is null when using automatic signing.
Something else is missing here but this fixes the issue for now.
The filename is returned by GetProvisioningData. It's explicitly set to null if bAutomaticSigning is true. Later on we attempt to build the path using that value which blows up if it's null. Interestingly we check for null later on but we don't do the check before trying to construct the FileReference. Added that check.

#jira ue-73506
#rb peter.sauerbrei

[CL 6621486 by Matt Collins in Dev-Rendering branch]