Fixed a bug in ARecastNavMesh::GetRandomReachablePointInRadius that resulted in erronous behavior for big(ish) Radius values.
The bug resulted from us using a strange extent ({Radius,Radius,Radius}) when projecting the initial location to navmesh. I've used a lot more reasonable DefaultQueryExtent for this purpose with Z being set to BIG_NUMBER. When a big Radius value was used the underlying recast code was expected to query a lot of polygons, but there's a hard limit of 128 polygons to process in multiple places in Recast lib (for performance reasons). These two facts combined resulted in finding awkward query starting point when using large Radius. The new approach first projects the query origin 'down' taking advantage of the assumption that the querier location is (more or less) on navmesh, since the whole query is about reachable locations.

#jira UE-66058
#review-6665296 @Yoan.StAmant
#rb Yoan.StAmant

[CL 6674755 by Mieszko Zielinski in Dev-Framework branch]