Sequencer: Fix autotracking.
On pre edit change, the changed property is checked against CanKey. This didn't work for structs, where the property chain would be something like DepthOfField->Focal Distance. A fix was made in 3729216 to support structs, but this broke generic properties because it would check CanKey with the property chain, LightComponent->Intensity.

The fix here is to find the keyable property and then use the remaining property path. For example, for intensity, the property path will be Intensity. For Focal Distance, the property path will be DepthOfField->Focal Distance.

#jira UE-74657
#rb mike.zyracki

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: max.chen
#ROBOMERGE-SOURCE: CL 6601177 via CL 6602290 via CL 6602390
#ROBOMERGE-BOT: ENGINE (Main -> Dev-Networking) (v365-6733468)

[CL 6747703 by max chen in Dev-Networking branch]