Generating stable shader info is now configurable.
Shader code library was always emitting "<shadermodel>.scl.csv" files while cooking. This output data is only needed for users (or automated jobs) that update the PSO cache. With this change the configuration value "NeedsShaderStableKeys" needs to be set to true for platforms that use PSO (e.g. AndroidEngine.ini and  IOSEngine.ini).

This saves diskspace and around 90 seconds on a iterative cook on a large project.

#rb Dmitriy.Dyomin

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: johan.berg
#ROBOMERGE-SOURCE: CL 6604300 via CL 6604799 via CL 6604833
#ROBOMERGE-BOT: ENGINE (Main -> Dev-Networking) (v365-6733468)

[CL 6754674 by johan berg in Dev-Networking branch]