#jira UE-75732 - Undo actions are blocked if the user duplicates an added component while naming it.
Before duplicating, clear the application keyboard focus to ensure that any component being renamed lost the focus and commit their text first. The duplicate operation triggers a regeneration of the tree view, releasing the row widgets. If one row had the keyboard focus (editing the component name), the 'weak focus path' maintained by Slate gets truncated (they are weak pointers, so this is fine) but when the focus changed, Slate could not pin the row widget that had the focus and therefore, the 'focus lost' event was never handled by the now deleted widget. Since the 'rename' transaction are performed on focus lost, the rename was not taken into account. For 'create + give initial name' transaction, the responsibility to end the create transaction was passed to the row item, but the transaction would not end because on focus lost was never called and the rename never ended. In that case, any further transactions were appended to the 'create' transaction until the node item was released. (Like deleting the component).

#rb Marc.Audy, Chris.Gagnon
#lockdown cristina.riveron

#ROBOMERGE-OWNER: lina.halper
#ROBOMERGE-AUTHOR: patrick.laflamme
#ROBOMERGE-SOURCE: CL 6953440 in //UE4/Release-4.22/... via CL 6953442
#ROBOMERGE-BOT: ANIM (Main -> Dev-Anim) (v367-6836689)

[CL 7084299 by patrick laflamme in Dev-Anim branch]