#ue4 - UpdateOverlaps() does nothing until after BeginPlay() has been called on the Actor. Fixed inconsistent ordering of UpdateOverlaps() relative to BeginPlay() in the cases of server / standalone SpawnActor, networked replicated actor, actor streamed as part of a level. BeginPlay() now always precedes UpdateOverlaps(), which was the previous behavior in standalone or on the server when spawning or as part of persistent map load.
Added UpdateOverlapsMethodDuringLevelStreaming to Actor, which allows you to disable the entire UpdateOverlaps() query run after BeginPlay(), or make it conditional based on mobility (skip for static objects). There is also an option to defer to the default specified in INI, which allows each class to have a separate default that is changeable also for engine classes without modifying engine source, and which makes it easily hotfixable.

The new default option is OnlyUpdateMovable, which means to runs overlaps only if the root component of the actor has Movable (non-static, non-stationary) mobility. Previous behavior was equivalent to AlwaysUpdate, however this can be horrible for streaming performance for many actor types.

#rb Marc.Audy, Andrew.Ladenberger

#ROBOMERGE-OWNER: lina.halper
#ROBOMERGE-AUTHOR: zak.middleton
#ROBOMERGE-SOURCE: CL 7107591 via CL 7110757 via CL 7110808
#ROBOMERGE-BOT: ANIM (Main -> Dev-Anim) (v367-6836689)

[CL 7136558 by zak middleton in Dev-Anim branch]