Make UFE/ProjectLauncher use the vanilla platform name for -platform when using Android cook flavors, rather than eg -platform=Android_ETC1 #jira UE-76146 #rb None
The old SetupTargetPlatforms() had a comment "// Need to tolerate cook platform names here, which UFE likes to pass in." but the data driven UnrealTargetPlatform Parse() no longer accepts non-canonical values for -platform. UFE always passes -targetplatform=Android and -cookflavor=ETC1 anyway. In the old code, the invalid -platform was ignored and the -targetplatform=Android was used. With this change the new code parses both and removes duplicates.

[FYI] Josh.Adams

#ROBOMERGE-SOURCE: CL 7138000 in //UE4/Release-4.23/...
#ROBOMERGE-BOT: RELEASE (Release-4.23 -> Main) (v367-6836689)

[CL 7138029 by jack porter in Main branch]