Fixed scene capture crash.
Rendering tiles using DrawTileMesh in a separate process without an accompanied main render can cause a crash in GetSceneTexturesUniformBuffer. This happens because translucency volume textures are not allocated until the first render in which they are used. However, the scene textures uniform buffer call attempts to dereference an array that is also not allocated until the textures are. This is incorrect. Instead, the call should return null which is then handled correctly by the uniform buffer call. Tile mesh rendering always uses the unlit mode, so it is never required to actually have these textures allocated.

#rb none
#jira UE-76507

#ROBOMERGE-OWNER: lina.halper
#ROBOMERGE-AUTHOR: zach.bethel
#ROBOMERGE-SOURCE: CL 7309942 in //UE4/Release-4.23/... via CL 7309958
#ROBOMERGE-BOT: ANIM (Main -> Dev-Anim) (v371-7306989)

[CL 7359134 by zach bethel in Dev-Anim branch]