#jira UE-78400 #jira UE-78464
Since we are compiling two architectures for HoloLens is causing the second architecture to overwrite the dependency list for the first file. Save them in different locations.

Originally we had a filter to explicitly copy the necessary files, but the filter was too wide. Now that we have this fix, removing the explicit copy filter.

#rb Ben.Marsh

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: joe.conley
#ROBOMERGE-SOURCE: CL 7846892 in //UE4/Release-4.23/... via CL 7846932
#ROBOMERGE-BOT: ENGINE (Main -> Dev-Networking) (v389-7813075)

[CL 7914339 by joe conley in Dev-Networking branch]