This changelist includes several fixes for issues related to or reported in the above ticket.
Fixes a case where a party/leader os rejected attempting to join a creative server they do not have a reservation as a leader on, but that one or more of the party members had an exiting reservation.
The code now checks that the existing reservations for those players aren't incompatible allowing the reservation only if all the members with existing reservations on the same team, and can the new members fit on that team.
(can be caused if that members fails to flush into the server on a previous join)

Fixes a case where a change of party leader would result in broken reservations being left on the beacon. 2 reservations with the same leader, and one in one of those the leader would not be a member.
This would create a situation where that leaser, or (baring the above fix) anyone in a party with them wouldn't be able to join for the lifetime of the server.

Fixes a case where party leaders who have a reservation on the server would be rejected attempting to rejoin (or join the case of a player in the lobby having been promoted to lead) unless their new reservation exactly matched the previous one in terms of party members (which never happens for a leader JIP). The code now also allows them to rejoin if the newly requested reservation is a subset of their existing reservation.

[REVIEW] [at]Mark.Fitt,  [at]Billy.McCarroll


#ROBOMERGE-SOURCE: CL 8062380 via CL 8062382 via CL 8062387 via CL 8092567
#ROBOMERGE-BOT: (v401-8057353)

[CL 8092690 by chris applegarth in Main branch]