The 'Audit Asset' dialog will now correctly refresh it's display of Chunks for the 'Editor' platform.
#rb Sebastian.Nordgren

AssetManager: Made sure that we clear CachedChunkMap before updating it so that ChunkIDs that have vbeen deleted during the editing session are properly discarded.
SAssetAuditBrowser: Instead of just refreshing the database for UAssetManager we refresh the data for the ManagerEditorModule, to make sure that it recaches the ChunkID's. This needs to be done once the dialog has been built as it can invoke callbacks to the SAssetAuditBrowser itself. This will result in the panel having up to date data when it is first displayed rather than requiring the user to press refresh immediately after opening it.

Note that the panel will not give sizes for chunks if the project is not set to generate chunked packfiles due to an optimization in CL 5672140, if that should be fixed or if we just prevent the chunks from being shown if the project will not create them is something for further consideration.

[CL 8111947 by paul chipchase in Dev-Core branch]