UMovieSceneSequencePlayer now relies on the external owner to call Stop() when the player is being torn down. BeginDestroy() is too late in the lifecycle (called immediately before deletion by GC) for UMovieSceneSequencePlayer to call Stop() on itself.
Stop has many potential side effects as it may try to restore state of external actors that this player is affecting. It is not safe to call this during BeginDestroy as the validity of object pointers is very turbulent during a GC pass. Now, the Level Sequence Actor (and ActorSequenceComponent) detect EndPlay() and call Stop (which is a noop if it's already stopped). EndPlay() is called when the original call to Destroy(...) goes out to tear down the owning actor at which point the validity of object pointers is still valid as we don't get GC'd until after EndPlay is called on destroyed items.

This fixes an issue where during a seamless level transition, Stop() would sometimes fail to disable Cinematic Mode on the Player Controller (which survives the transition) as the World would have been destroyed at that point and we use the world to get the local player controller.

#jira UE-79481
#rb Max.Chen

[CL 8386282 by Matt Hoffman in Dev-Editor branch]