UECORE-482: Crash reporting in external process.
With this change an additional mode is added to crash reporter client, that allows it to run as a monitoring child process. Additionally stack walking and symbol resolution is moved to this new process on Windows. To support this several changes in stack walking had to be made to allow external processes to be queried. All other platforms remain unchanged.

#jira UECORE-482, UE-72081, UE-75336
#rb ben.marsh, brandon.schaefer, patrick.laflamme

[CL 8395425 by Johan Berg in Dev-Core branch]