Use memoryless depth buffer on mobile platforms which support it.
Replaced the unused TexCreate_BiasNormalMap texture creation flag with a new TexCreate_Memoryless flag which can be used to request render targets which only exist in tile memory on TBDR GPUs. The mobile renderer sets this flag when it knows that the contents of the depth buffer will not be read.

Implemented support for TexCreate_Memoryless on Metal (copied from CL 8129626) and Vulkan. On Vulkan, memoryless mode is requested by adding VK_IMAGE_USAGE_TRANSIENT_ATTACHMENT_BIT to the image usage flags, and allocating the target from a memory type which supports VK_MEMORY_PROPERTY_LAZILY_ALLOCATED_BIT. If no such memory type is exposed by the device, it means it doesn't support memoryless targets, so we fall back to regular allocations.

#rb dmitriy.dyomin matt.collins

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: mihnea.balta
#ROBOMERGE-SOURCE: CL 8633327 via CL 8633466 via CL 8633469
#ROBOMERGE-BOT: ENGINE (Main -> Dev-Networking) (v422-8689730)

[CL 8738206 by mihnea balta in Dev-Networking branch]