Blueprint: Added support for generated functions within Blueprint Graphs
Generated function graphs can be created as part of the compilation process and are ephemeral (only lasting the duration of the compile), but will still be used to generate bytecode and UFunctions on the final class layout. External clients can hook into this process by binding to UBlueprint::GenerateFunctionGraphsEvent and spawning the necessary intermediate graphs on the compiler context using FKismetCompiler::SpawnIntermediateFunctionGraph. GenerateFunctionGraphsEvent is a dynamic delegate, meaning that all bindings will be serialized into the package, enabling composition workflows within a single package.

Generated function graphs are also visible as an intermediate build product with this Blueprint developer setting enabled.

#rb Dan.OConner

[CL 9068567 by Andrew Rodham in Dev-Editor branch]