Sequencer: Changed Transform origins to be implemented as shared global data rather to remove the need to Initialize transform tracks
Transform tracks are heavily used, whereas transform origins are not. As such we want to remove the ::Initialize call that is unnecessary for the majority of transform tracks. We now compute the sequence transform origin once at the start of the frame, and use that for the rest of the evaluation.

#rb Max.Chen, Matt.Hoffman, Mike.Zyracki

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: andrew.rodham
#ROBOMERGE-SOURCE: CL 9159615 via CL 9160214 via CL 9160238 via CL 9160259
#ROBOMERGE-BOT: ENGINE (Main -> Dev-Networking) (v443-9013191)

[CL 9182708 by andrew rodham in Dev-Networking branch]