Sequencer: Prevent pre-animated state being restored on spawnables that are already destroyed.
This change introduces a new flag to evaluation templates: whether to prioritize tear down or not. Normally tear down strictly occurs in the reverse order of setup, however this is not desireable for spawned objects since we end up restoring a bunch of state such as transforms, mobility, attachments etc on objects that are going to be destroyed that same evaluation. Spawn tracks now specify this prioritized tear down which means that the object will always be destroyed first, and any subsequent calls to RestorePreAnimatedState will early out because the object is now pending destruction.

#rb Max.Chen, Mike.Zyracki, Matt.Hoffman

#ROBOMERGE-OWNER: ryan.gerleve
#ROBOMERGE-AUTHOR: andrew.rodham
#ROBOMERGE-SOURCE: CL 9159674 via CL 9160216 via CL 9160241 via CL 9160261
#ROBOMERGE-BOT: ENGINE (Main -> Dev-Networking) (v443-9013191)

[CL 9182710 by andrew rodham in Dev-Networking branch]