Add target for editor crash report client
There are some operations in the editor (package and stage) that needs to build Crash report client, but when we run it in monitor mode the binary is locked failing the build. This change adds a new target (basically a copy of the crash report client) for the editor. This also allows us to add specific dependencies for the editor client (such as disaster recovery) which should not be shipped with the regular version.

#rb ben.marsh

[CL 9297523 by Johan Berg in Dev-Core branch]