From Sebastien Hillaire:
Fixed extinction clamp to only be used where it is needed.
Avoid black water while allowing opticaldepth=0 and thus transmittance=0 when absorption+scattering=0.

This is a follow up to 9425005. Transmittance was still not 0 when abs=scatt=0, it is just that the threshold is small so very low opaticalDepth and black will happen on mobile. Now All platforms have same consistent behavior.

#rb Krzysztof.N


#ROBOMERGE-SOURCE: CL 9426261 via CL 9426262 via CL 9426263 via CL 9426410 via CL 9426431
#ROBOMERGE-BOT: (v486-9426253)

[CL 9426451 by matt kuhlenschmidt in Main branch]