Removing some code in-preparation of thread-safety fix
We shouldn't care if *all* custom versions are present in any particular package that is being loaded. The whole point of custom versions is to allow different systems to versions and upgrade data independently. Everything that should matter is if all of the relevant custom versions are up to date or not.

#rb steve-robb

[CL 9566340 by Johan Torp in Main branch]