Fix crash when editing and saving a default material
When FMaterialEditor::UpdateOriginalMaterial() duplicates the original material (due to hitting save in the material editor), the PostLoad() callback inside UMaterial recompiles the shader. The bUsedAsSpecialEngineMaterial flag hasn't been set yet on the copy (it's set after StaticDuplicateObject() returns), so the copy may not be usable with all the object types that the original material was used for. When this is the case, e.g. when the material is assigned to a spline mesh, we trigger an assert in FMaterial::GetShaderPipeline().

FMaterialEditor::UpdateOriginalMaterial() uses UMaterial::ForceNoCompilationInPostLoad() to disable compiling during duplication, but the flag that's set by that function was not actually used anywhere. The check for GAllowCompilationInPostLoad seems to have been accidentally removed in CL 2044510.

#jira UE-80327
#rb rolando.caloca

[CL 9614424 by mihnea balta in 4.24 branch]