First pass for creating a manifest describing all BulkData encountered during a cook for ZenCreator so that it can eventually be allowing it to be used by the IoDispatcher. Currently it is ignoring the existence of optional and duplicated non optional bulkdata which can be added later.
When cooking the package we now create a manifest of all potential bulkdata access patterns to the output ubulk files so that ZenCreator knows aobut them.

At the moment BulkData does not have access to the FGlobalNameMap and so using the PackageNameIndex to generate it's FIoChunkId will not work, so currently I am just setting it to zero. In future work BulkDatas will create this token during the cooking phase to reduce cpu work at load time and ideally gain access to FGlobalNameMap so that it can use it in it's generation.

Currently the new BulkData flag BULKDATA_CookedForIoDispatcher serves two purposes, firstly it allows the old BulkData to check if it is loading up cooked data with the new index, or if it's loading older bulkdata from a uasset, secondly at runtime for the new BulkData system (in a later submit) we can use it to decide if we load from cooked loose files or the IODispatcher system. These use cases will probably change over the next few weeks as I clean up the code.

#rb PJ Kack

[CL 9630644 by paul chipchase in Dev-Core branch]