Merging 9546470 from //UE4/Dev-CharacterTech to //UE4/Release-4.24
For Pixel Streaming added a client-side AFK system which will disconnect inactive users. It will first warn the user after a period of inactivty that they are to be disconnected and a countdown will be shown. If the user does not click then they will be disconnected.
This is configurated in app.js as follows:
afk.enabled: Set this to true to activate the AFK system; default false.
afk.warnTimeout: Set this to the number of seconds of inactivity before the warning overlay is displayed; default 120 seconds.
afk.closeTimeout: Set this to the number of seconds the warning overlay will be displayed before the user is disconnected (unless they click); default 10 seconds.

#jira FORT-207270
#rb Alistair.White
[RN] Added a client-side AFK system which will disconnect inactive users.

#ROBOMERGE-SOURCE: CL 9688071 in //UE4/Release-4.24/...
#ROBOMERGE-BOT: RELEASE (Release-4.24 -> Main) (v538-9681622)

[CL 9688076 by alistair white in Main branch]