Copied CL# 9693447
#jira UE-81409
Bug fixed: Renaming ULayers did not update their internal UActors nor the ViewHiddenLayers that contained them if it was called with a reference to the Layer->LayerName, e.g. RenameLayer(Layer->LayerName, NewLayerName). This created lots of side effects with UActors and HiddenLayers, that after the layer rename pointed to no-longer-existing Layer names.
#rb codereview chris.gagnon

#ROBOMERGE-SOURCE: CL 9693857 in //UE4/Release-4.24/...
#ROBOMERGE-BOT: RELEASE (Release-4.24 -> Main) (v538-9681622)

[CL 9693862 by gines hidalgo in Main branch]