Fix for stack-use-after-scope error caught by Address Sanitizer.
In FSlateUser::SynthesizeCursorMoveIfNeeded, the SyntheticCursorMoveEvent can be constructed with an empty set of pressed mouse buttons.  The empty set was constructed as a parameter, and falling out of scope.  The problem is that FPointerEvent keeps a reference to that parameter, and accesses it later.  The fix is to construct an empty set prior to creating the SyntheticCursorMoveEvent, and just reference it in the constructor parameters, if necessary.

[REVIEW] [at]thomas.ross
#rnx
#lockdown Chris.Bunner


#ROBOMERGE-SOURCE: CL 9732309 via CL 9732312 via CL 9732313 via CL 9732314
#ROBOMERGE-BOT: (v539-9700858)

[CL 9732316 by will damon in Main branch]