Editor: Fixed reallocation of auto-reimport monitors causinginvalid delegate bindings
FContentDirectoryMonitor is a self-referential type since it contains an FFileCache which is also self-referential. This means that reallocation of the monitors can result in invalid delegates being bound for the file monitors. This has been addressed by heap-allocating the monitors (so they are never moved or copied themselves), along with some other trivial changes to bring this code more up-to-date with modern UE4 coding standards.
This is largely in response to github pr 5906 but is a more thorough solution.

#jira UE-75648
#rb None
#5906

#ROBOMERGE-SOURCE: CL 9744103 in //UE4/Release-4.24/...
#ROBOMERGE-BOT: RELEASE (Release-4.24 -> Main) (v542-9736015)

[CL 9744104 by andrew rodham in Main branch]