A (temporary) support for stationnary/static lighting for hair rendering.
This allows to have some lighting onto hair cluster/primitives. However static geometry won't cast any occlusion/shadow onto hair, as currently there is no setup for computing pre-shadow for hair. The main limitation is that hair are neither visible inot the main pass nor into the shadow (both are handled with a custom manner). The side effect of this is that it is difficult to render pre-shadow for it. This needs to be address later one.

#rb none
#jira UE-82298

#ROBOMERGE-SOURCE: CL 9823374 in //UE4/Release-4.24/...
#ROBOMERGE-BOT: RELEASE (Release-4.24 -> Main) (v546-9757112)

[CL 9823381 by charles derousiers in Main branch]