Allow game/editor to continue execution earlier after a crash.
A previous change moved the signal to the game/editor that it's okay to continue to after the crash report client was completely done with sending and resolving callstacks, because it was assumed that there was a syncronization problem. However that proved to be another issue, so moving the signal back to where it was originally. This should make the editor only "freeze" a short time, while necessary data is collected.

#jira UE-82333
#rb pj.kack

(ushell-p4-cherrypick of 9868282 by Johan.Berg)

#ROBOMERGE-SOURCE: CL 9868804 in //UE4/Release-4.24/...
#ROBOMERGE-BOT: RELEASE (Release-4.24 -> Main) (v548-9842178)

[CL 9868810 by johan berg in Main branch]