BulkData2 can now use the FIoDispatcher when it is enabled for the given platform.
NOTE: This is still disabled by default.

Inline data does not currently use the FIoDispatcher (I am not sure if we want to support reloading the inline data at runtime, we currently have no active use cases and it would inflate the toc) so uses the same fallback members in the union as we use when not running with the FIoDispatcher. Using a union allows us to keep the memory footprint very small. BulkDataSize was moved from the FileToken system to the Fallback members so that inline data does not need to register a token which prevents pointless memory allocations.
BulkData can be polled by calling ::IsUsingIODispatcher to see if it is using the FIoDispatcher or not.
NOTE: If the Zenloader is in use then BulkData will no longer need to try and access a FLinkerLoad and will not try to find any filenames.

Implemented a FIoDispatcher specific implementation FBulkDataIoDispatcherRequest, this mirrors the IPlatformFIle::OpenAsyncRead api as it moved the IO request to a AsyncTask and implements all of the same functionality and provides the same callback interface although that callback expects a IAsyncReadRequest pointer which it cannot prrovide. It also fakes the ability to use user supplied memory and issue multiple FIoDispatcher requests and have them returns in the same continious memory buffer. These features will probably be moved to FIoDispatcher and could be removed from this code over time.

The FIoDispatcher is passed to the BulkData system at the point of creation by FAsyncLoadingThread2Impl and held in a static member.
Changing ENABLE_IO_DISPATCHER to 0 will cause the system to read the loose files and not use IoDispatcher even if the data exists.
Removed BulkDataSize from the FileTokenSystem.
Removed the == operator from IBulkDataIORequest as it is no longer used, some methods had const removed as the FIoDispatcher implementation would not be const.

The old 16bit ChunkID has been replaced by a proper FIoChunkID which is generated by using info from the GlobalName map when cooking the bulkdata (BulkData will save out dummy values that are filled in later when actually saving the package, such as offsets etc, the ChunkID is now one of these values)
Previously I was writing out the uint16 ChunkID during saving of BulkData but it should've been when saving AND cooking to avoid writing it to the uasset or DDC, in order to fix this I had to rename BULKDATA_CookedForIoDispatcher to BULKDATA_BadDataVersion so I can specifically check for it on load and discard the value. Once things stablize I can increase the DDC keys for the data types that write BulkData to the DDC and remove BULKDATA_BadDataVersion.

::CreateStreamingRequestForRange now takes for the full range of reads requested rather than just the start and end in the form of a TArray (BulkDataRangeArray). This is so we can make a unique request for each read when using the FIoDispatcher and keeps the old and new code paths the same in calling code.

Known issues: Optional data will still not function correctly and Virtual Textures are still using older code paths and require the loose ubulk files still. The IOS memory mapping systems have no actual support. This has added more code paths with BULKDATA_NOT_IMPLEMENTED_FOR_RUNTIME which will need to be handled properly soon (either fixed up or log warnings for improper use)

#rb Per.Larsson

[CL 9879748 by paul chipchase in Dev-Core branch]