Fix blocking code in NetworkChangedManager
# Summary
Since the HTTP url open connection is a blocking call. It needs to be run on a seperate thread. The call .get on a future is blocking though, meaning we were still running a blocking call on the main looper. This updates that code to instead add a timeout runnable to the local handler, to cancel the executor if it hasn't completed in the allotted time.

Added some tweaks to match comments made by [at]Pete.Procopio
 - Refactor INSTANCE to be lower case
 - Tweak construction of Logger to match current pattern
 - Copied over correct NetworkChangedManager to remove System static call that was accidently included in the NativeWrapper

[CODEREVIEW] [at]Zhao.Zhong [at]Chris.Babcock [at]Pete.Procopio [at]trevor.stone

# Test Plan
1. Built and ran app with UeDevelopmentDebug
2. app launched
3. signed in
4. verified that could start download
5. Turned off wifi
6. Verified that saw network error
7. Turned wifi back on
8. Verified that network error went away automatically



#ROBOMERGE-SOURCE: CL 10073998 via CL 10074011 via CL 10074013 via CL 10074017
#ROBOMERGE-BOT: (v574-10069753)

[CL 10075107 by ben temple in Main branch]